+++ api/src/glfs.c
@@ -845,11 +845,6 @@ pub_glfs_new(const char *volname)
     }
 
 label:
-    /*
-     * Do this as soon as possible in case something else depends on
-     * pool allocations.
-     */
-    mem_pools_init();
 
     fs = glfs_new_fs(volname);
     if (!fs)
@@ -906,9 +901,6 @@ out:
         if (fs) {
             glfs_fini(fs);
             fs = NULL;
-        } else {
-            /* glfs_fini() calls mem_pools_fini() too */
-            mem_pools_fini();
         }
     }
 
@@ -1400,12 +1392,6 @@ pub_glfs_fini(struct glfs *fs)
 
 free_fs:
     glfs_free_from_ctx(fs);
-
-    /*
-     * Do this as late as possible in case anything else has (or
-     * grows) a dependency on mem-pool allocations.
-     */
-    mem_pools_fini();
 
 fail:
     if (!ret)
