+++ libglusterfs/src/event-poll.c
@@ -18,7 +18,7 @@
 
 #include "glusterfs/logging.h"
 #include "glusterfs/gf-event.h"
-#include "glusterfs/mem-pool.h"
+#include "glusterfs/memory.h"
 #include "glusterfs/common-utils.h"
 #include "glusterfs/syscall.h"
 #include "glusterfs/libglusterfs-messages.h"
