+++ libglusterfs/src/gidcache.c
@@ -9,7 +9,7 @@
 */
 
 #include "glusterfs/gidcache.h"
-#include "glusterfs/mem-pool.h"
+#include "glusterfs/memory.h"
 #include "glusterfs/common-utils.h"
 
 /*
