+++ libglusterfs/src/glusterfs/xlator.h
@@ -1075,8 +1075,6 @@ copy_opts_to_child(xlator_t *src, xlator_t *dst, char *glob);
 
 int
 glusterfs_delete_volfile_checksum(glusterfs_ctx_t *ctx, const char *volfile_id);
-int
-xlator_memrec_free(xlator_t *xl);
 
 void
 xlator_mem_cleanup(xlator_t *this);
