+++ libglusterfs/src/graph.c
@@ -30,7 +30,7 @@
 #include "glusterfs/libglusterfs-messages.h"  // for LG_MSG_GRAPH_ERROR, LG_...
 #include "glusterfs/list.h"                   // for list_add, list_del_init
 #include "glusterfs/logging.h"                // for gf_msg, GF_LOG_ERROR
-#include "glusterfs/mem-pool.h"               // for GF_FREE, gf_strdup, GF_...
+#include "glusterfs/memory.h"                 // for GF_FREE, gf_strdup, GF_...
 #include "glusterfs/mem-types.h"              // for gf_common_mt_xlator_list_t
 #include "glusterfs/options.h"                // for xlator_tree_reconfigure
 #include "glusterfs/syscall.h"                // for sys_close, sys_stat
