+++ libglusterfs/src/rbthash.c
@@ -11,7 +11,7 @@
 #include "glusterfs/rbthash.h"
 #include "rb.h"
 #include "glusterfs/locking.h"
-#include "glusterfs/mem-pool.h"
+#include "glusterfs/memory.h"
 #include "glusterfs/logging.h"
 #include "glusterfs/libglusterfs-messages.h"
