+++ libglusterfs/src/rot-buffs.c
@@ -11,7 +11,7 @@
 #include <math.h>
 
 #include "glusterfs/mem-types.h"
-#include "glusterfs/mem-pool.h"
+#include "glusterfs/memory.h"
 
 #include "glusterfs/rot-buffs.h"
