+++ rpc/rpc-lib/src/rpc-clnt.c
@@ -16,7 +16,7 @@
 #include "xdr-rpcclnt.h"
 #include "rpc-transport.h"
 #include "protocol-common.h"
-#include <glusterfs/mem-pool.h>
+#include <glusterfs/memory.h>
 #include "xdr-rpc.h"
 #include "rpc-common-xdr.h"
