+++ rpc/rpc-lib/src/rpc-drc.c
@@ -14,7 +14,7 @@
 #endif
 #include <glusterfs/locking.h>
 #include <glusterfs/statedump.h>
-#include <glusterfs/mem-pool.h>
+#include <glusterfs/memory.h>
 
 #include <netinet/in.h>
 #include <unistd.h>
