+++ xlators/features/changelog/lib/src/gf-changelog.c
@@ -200,8 +200,6 @@ gf_changelog_cleanup_this(xlator_t *this)
 
     this->private = NULL;
     this->ctx = NULL;
-
-    mem_pools_fini();
 }
 
 static int
@@ -235,12 +233,7 @@ error_return:
 static int
 gf_changelog_init_primary()
 {
-    int ret = 0;
-
-    ret = gf_changelog_init_context();
-    mem_pools_init();
-
-    return ret;
+    return gf_changelog_init_context();
 }
 
 /* TODO: cleanup clnt/svc on failure */
