+++ xlators/nfs/server/src/acl3.c
@@ -13,7 +13,7 @@
 #include <glusterfs/dict.h>
 #include <glusterfs/xlator.h>
 #include "nfs.h"
-#include <glusterfs/mem-pool.h>
+#include <glusterfs/memory.h>
 #include <glusterfs/logging.h>
 #include "nfs-fops.h"
 #include "nfs3.h"
