+++ xlators/nfs/server/src/mount3-auth.c
@@ -38,7 +38,7 @@
 #include "mount3-auth.h"
 #include "exports.h"
 #include "netgroups.h"
-#include <glusterfs/mem-pool.h>
+#include <glusterfs/memory.h>
 #include "nfs-messages.h"
 
 /**
