+++ xlators/nfs/server/src/mount3udp_svc.c
@@ -10,7 +10,7 @@
 
 #include "xdr-nfs3.h"
 #include <glusterfs/logging.h>
-#include <glusterfs/mem-pool.h>
+#include <glusterfs/memory.h>
 #include "nfs-mem-types.h"
 #include "nfs-messages.h"
 #include "mount3.h"
