+++ xlators/nfs/server/src/nfs.c
@@ -17,7 +17,7 @@
 #include <glusterfs/dict.h>
 #include <glusterfs/xlator.h>
 #include "nfs.h"
-#include <glusterfs/mem-pool.h>
+#include <glusterfs/memory.h>
 #include <glusterfs/logging.h>
 #include "nfs-fops.h"
 #include "mount3.h"
